Update of the learning algorithm
Previously I used only two actions. One was for braking and accelerating and the other one for steering. The problem was that you can't reinforce the opposite if the car did nothing because the opposite would be -1 and 1 at the same time. Therefore I added 4 actions. One for every possible action and the agent takes the most likely one. If one action leaded to a negative reward I could just increase the likelihood for the other 3 actions.